particle updater refactor
